server/backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
httpx>=0.27.0,<1.0.0
kafka-python>=2.0.2,<3.0.0
pytest>=8.0.0,<9.0.0
